+++ jib-maven-plugin/scripts/prepare_release.sh
@@ -39,8 +39,8 @@ if [[ $(git status -uno --porcelain) ]]; then
     Die 'There are uncommitted changes.'
 fi
 
-# Runs integration tests.
-./gradlew jib-maven-plugin:integrationTest --info --stacktrace
+# Runs checks integration tests.
+./gradlew jib-maven-plugin:check jib-maven-plugin:integrationTest --info --stacktrace
 
 # Checks out a new branch for this version release (eg. 1.5.7).
 BRANCH=maven_release_v${VERSION}